Refactor DirectoryWatcher/FileCache to limit depth
I decided that a boolean toggle for recursive watching was not
sufficiently powerful, so I switched to adding a depth parameter. While
I do think that recursive or not is by far the most common use case,
given that this is intended to be a general purpose library, I think
that it's better to be explicit about depth. The maintenance cost should
be pretty small.

I left the register methods that have the recursive parameter for
convenience.