Improve code converter
There were a couple of bad artifacts in the generated scala.js scaladoc.
Mostly it was the case that {@link foo#bar} was getting truncated to
[[foo.ba]] and some hyperlinks were not being correctly converted from
<a href="$LINK"( target=_blank)?>$NAME</a> to [[$LINK $NAME]]. God help
me the next time I have to read the StripArgs regex.

I also made it slightly more debuggable in the repl by reworking the
Converter object.