Add compile errors for with*Sync misuse
I noticed that some tests would occasionally print errors without
actually failing. This was because the body of, say, a withTempFileSync
actually returned a Future rather than a value. To prevent this, I more
or less copied the <:!< trait out of shapeless and now require that
thunk passed into the with*Sync methods returns a value that does not
inherit from Future[_].