Bump timeout for all builds
Occasionally the osx addmany test fails on travis. I think that this is
because of timeouts, not because of logic errors. It turns out that my
test reporting was not quite right because I was calling andThen after
the cache was closed. I wanted to know if events were still being
generated after the timeout, so I needed to move it to after the last
flatMap.

I also uncovered an issue for scala.js (especially on windows), that the
test execution context did not necessarily run callbacks in order. I
fixed this by adding a queue of runnables and running them in order.