Update locking system
Rather than directly using a ReentrantLock, I introduce a Lock class
whose lock method is interruptible. I made this change because sometimes
I ran into problems where one thread would interrupt a thread that was
blocked on lock acquisition and because it was non-interruptible, the
thread would not stop and eventually a downstream error would crop up.
To avoid this, the new Lock is interruptible and returns true if it
acquires the lock, false otherwise.

Most of this commit is just white space because I had to move a lot of
code inside of if (lock.lock()) { ... }. Indeed, the commit is far more
readable with the '-w' flag.