Fix miscellaneous issues in DirectoryWatchers
I discovered two big issues:
1) NioDirectoryWatcher could have issues if the callbackExecutor got
   backed up and threw a RejectedExecutionException.
2) Neither the NioDirectoryWatcher nor the AppleDirectoryWatcher
   properly handled dynamically changing the depth of a registered
   directory.

For (1), I added an additional overflow handler. If the callbackExecutor
rejects the callback, we will block the loop thread until the
callbackExecutor finishes a callback. We then run an overflow handler
and move on with our lives.

For (2), I just needed to add some logic for properly tracking the
registered paths and checking for overlaps when register is called.  I
added some tests to DirectoryWatcherTest and NioDirectoryWatcherTest to
address (2). In an ideal world, I'd consolidate the three
DirectoryWatcher tests since they have a lot of overlap and could
probably be factored out, I am not going to do that right now. As a
result, there are two identical copy/paste tests in
NioDirectoryWatcherTest and DirectoryWatcherTest.