Rework concurrency
I continued to be dogged by spurious test failures and realized that the
locks were making things hard to reason about. To get around this, I
reworked a number of classes to move work onto an executor. The
advantage to this approach is that we can ensure that everything on the
executor happens on the same thread. The cost of this approach is that
it makes the code a bit more verbose because of all of the runnables
that I have to create and that there is a corresponding increase in the
jar size (from O(1.35Kb) to O(1.52Kb)). The nice thing is that I don't
see any more spurious failures after this change! It is crucial that the
Executor.run method not submit the runnable to the executor service if
it's already running on the executor. I updated the ThreadFactory class
to keep track of the threads that it created so that the executor can
check whether it's already on an executor thread when it enters the run
method.

In order to control which thread the callbacks are invoked on, I add a
copy method to the Executor which allows a class to pass down the
Executor to a child class. When the parent closes the child, the child
may close the executor, but it will not close the executor until the
parent (which created the executor) closes it.

I also made it possible to run a blocking operation on the execution
context. This is so that the public apis, e.g. register, will block, but
the actual work will occur on the executor thread.

This change would not be nearly so ugly with lambdas, but I'm sticking
with java 7 for now. Long term, I may branch to java 8 so that I can use
lambdas in the implementation. At that point, I could publish java 7
compatible binaries with retrolambda. Hmm...