Add cast methods to Either
I wanted to be able to cast an Either<L, R> to an Either<+L, R> or
Either<L, +R>. This is mildly annoying since java doesn't declaring
variance for classes.

I wanted to be able to call Executor.block and only catch specific
exceptions. It is relatively straightforward to implement by passing in
the exception class to a second block method. I also added an additional
covariant type parameter to the Either left and right methods so that I
wouldn't have to cast subclass to their parent when constructing
Eithers.