Improve Executor.block
This is a bit cleaner. The queue.add method shouldn't throw so this is
easier to reason about. It is lame that I had to add the three type
parameters but java type inference is not nearly as powerful as scala
(where I probably wouldn't need a cast at all and the declaration of
Either.left and Either.right would require only two type parameters
because I could declare the covariance in the class declaration.