Return Either<IOException, Boolean> instead of boolean
It makes more sense to have the register methods of DirectoryWatcher
return an Either<IOException, Boolean> as opposed to a boolean (with no
thrown checked exception). The advantage to this approach over a checked
exception is that it makes it more obvious in other languages that
compile to jvm byte code that the error needs to be handled.