Allow Entry value to be null
I realized that we don't want everything to break if the cache cannot
successfully compute the value for the Directory.Entry. To avoid this, I
updated the Entry to include an optional exception member variable that
caches the IOException that was thrown attempting to compute the cache
value. I also removed direct access to the value and added getters for
the value and the exception.