Add optional filter to Directory
I want to be able to build a directory that filters some entries (i.e.
only includes directories). For now, I'm leaving this package private,
but I may make it public. I worry about adding filters to the file cache
because it makes it difficult to reason about when different paths are
registered with different filters.