Rework Directory to handle updates in a more formal way
It was somewhat error prone the way that I was returning the updates
when the addUpdate method was called. For example, it wasn't possible to
determine if an update was a delete or a create from the Entry<T>[]
returned by addUpdate. To fix this, I added an Update class that tracks
all of the different type of events. This makes it possible to more
correctly handle the events without additional file system polling.

I also renamed addUpdate to update because addUpdate was a horrible
name.

I added the MapOps class that provides a utility method for diffing the
contents of a map. This is really an implementation detail so I make the
class package private.