Add Options to directory watcher
I added code that makes the DirectoryWatcher poll a new directory and
create new events for any files found in the newly create directory.
The FileCache already was polling when the DirectoryWatcher reported new
directories so this polling was redundant. The redundant polling was
slow, so I made it an option in the DirectoryWatcher. I also reworked
the AppleDirectoryWatcher so that the flags and latency could be passed
in as options. This change allowed me to clean up the DirectoryWatcher
factory methods since I no longer had to add the apple directory watcher
flags and latency as parameters.