Fix depth bug
After refactoring NioDirectoryWatcher, the tests that depth was
correctly managed would occasionally fail O(1%) of the time. I finally
tracked it down to adding a subdirectory with a depth of -1. This would
cause a subsequent call to update for a path that possibly should have
been ignored to get added to that subdirectory in spite of the limit.