Rework NioDirectoryWatcher
I wanted to make it possible to register directories that may not exist
as well as monitor files in addition to directories. In order to do
this, I needed to be able to guarantee that we detect when the missing
directory is created. It was hard to do this with the existing
architecture, so I re-implemented the NioDirectoryWatcher entirely. Now,
it creates a tree of watched directories starting from the root path. If
a path is registered that doesn't exist, it will register all of the
directories that exist and are part of the prefix of the registered
directory. For example, if the path /foo/bar/baz is registered and only
/foo exists, the paths "/" and "/foo" will be registered. If "/foo/bar" is
created, then "/foo/bar" will be registered, but if "/foo/baz" is
registered and there is no registered directory with that prefix, then
"/foo/baz" will not be registered.

With this approach, it's much easier to reason about the state of the
watcher. Moreover, it facilitates registering files, because a file can
just be treated in the watcher as a missing path so that in the above
example if "/foo/bar/baz" is a file, then "/foo/bar/" will be registered
and detect updates to "/foo/bar/baz".

While making this change, I also factored out some functionality so that
the js NioDirectoryWatcher could share most of the registration logic.
To make this work, I add the IO functional interface which, in this
case, specifies how to register a path with the underlying watch service
(which for js is provided by the node file system api). Each
implementation must also create a way for invoking callbacks when
updates occur. For js, this is trivial because registration involves
passing the framework a calllback. For the jvm, we need to add a thread
for pulling events from the watch service. By moving this logic on the
jvm out of the base NioDirectoryWatcher, I was able to factor out most
of the common logic, which is a pain to get right. Moreover, because
they have different concurrency patterns, I was able to identify some
concurrency bugs that (theoretically) would affect the jvm in the js
tests.