Add an abstraction for ShutdownHooks
I wanted to ensure that a number of classes were closed at shutdown. The
main reason was that the FileEventsApi can segfault if the callback that
is passed in is garbage collected. This only happens at shutdown, but
did cause a number of segfaults (and associated error files) for me.

Since scala.js doesn't have shutdown hooks, I made an abstraction for
it that is a no-op on scala.js.