Close file cache on shutdown
When sbt exits, it doesn't close the file cache. The problem with this
is that in some case it can cause a segfault because the file events api
tries to call back into the jvm using a class that has been garbage
collected. This leads to a segfault. To avoid this, ensure that we close
the cache on shutdown.