Handle symlinks better in directory
The Directory didn't really handle symlink loops if they were present.
To fix this, I add some checks when adding a subdirectory that the
subdirectory is not a loop. The new semantics of the directory is that
if there is a symlink loop, the link that completes the loop will be
visible, but it will not be traversed.