Handle loops in SymlinkWatcher
The symlink tests for the FileCache would sometimes fail in the scala.js
environment. It turns out that the problem was that the SymlinkWatcher
was feeding loops back into the directory. This didn't happen very often
on the JVM because it was generally induced when the tests were run in
parallel and side effects from other tests influenced the state of the
directories when the callbacks were invoked.