Add FileCache support for monitoring files
Previously, the FileCache only supported monitoring directories.
Moreover, the directory needed to exist in order for registration to
succeed. With the recent changes that I've made to the DirectoryWatcher,
we can relax this restriction and allow the user to register directories
that do not yet exist as well as to register files directly.

After this change, the FileCache should be able to accurately handle the
case where a directory is recreated. It also should be able to handle
files that are directly registered with it.

I also make explicit the expected semantics of the list api of the
cache. The semantics are as follow:

* If the path parameter is a file
  -  If the file was registered directly with the cache, return a list
     containing the file entry
  -  If the file is a child of a directory that was registered with the
     cache and whose max depth covers the file, return a list containing
     the file entry, otherwise return an empty list
* If the path parameter is a directory
  -  If the directory was registered directly with the cache
    / If the max depth was -1, then return the directory
    / Otherwise return the files contained in the directory, up to the
      minimum of the max depth the directory was registered with and the
      max depth parameter passed into the method
  - If the directory is a child of a registered directory and is covered
    by the max depth of the parent, return the entries contained in the
    directory up to the minimum of the max depth of the parent minus the
    distance from the parent to the directory and the max depth
    parameter