Fix NioDirectoryWatcher watch management
There were a number of small issues in the NioDirectoryWatcher. One was
that if a directory was removed, we did not actually remove the watcher.
This was surprisingly not caught by the existing tests, but was a bug.
Another issue was that when the watcher encountered an error in
scala.js, we did not clean up the watcher.