Add support for negative directory depth
It turns out that it's convenient to be able to have a directory that
has negative depth. This directory can be interpreted as representing a
file rather than a directory. The semantics of listing this directory
should be the same as `ls.` When calling `ls` on a file, the file itself
is returned. When calling `ls` on a directory, the files contained in
the directory are returned, but not the directory itself.