Add native library garbage collection
There is no reliable way to unload a dll within a jvm session without
using reflection techniques that are illegal in java >= 9. This makes it
impossible to remove the temporary dll on Windows. Even on linux and
mac, it probably isn't the best idea to delete the temporary shared
library. Instead of deleting the file in the shutdown hook, I instead
create a file with the suffix `.delete`. I then add a simple garbage
collector that looks in the temporary library directory for all of the
files that are marked with delete and removes them.