Handle exceptions in executor
It was possible for the executor to die because of uncaught exceptions
in the runnables that it was invoking. To fix this, I refactor the
jvm executor implementation to submit a task to the thread that polls a
queue of Runnables. It tries to run the runnable, but catches all
uncaught Exceptions thrown by the runnable and logs them. To enable
quick shutdown of the executor even when there are many enqueued
runnables, I actually maintain a queue of Either<Integer, Runnable>. To
shutdown, we drain the queue and offer a poison pill, which is
Either.left of any integer value.

The motivation was that I found that the plugin didn't work in windows
because the FileCache would get stuck on Executor.block.