Simplify SymlinkWatcher
Now that the PathWatcher can monitor paths, I don't need to worry about
whether the input path is a directory or a path. This allows me to
remove a bunch of code from SymlinkWatcher and also increase code
coverage (since some of the branches in SymlinkWatcher were difficult to
hit).