Fix FileCache overflow handler
In the overflow case, the FileCache was not actually registering newly
found keys with the watch service. As a result, the test mainly worked
by accident because the deletion case also relied on overflows to
ensure that we detected the deleted files. I updated the test to ensure
that file events are detected in the newly created subdirectories.