Refactor FileCache into an interface
I am not a huge fan of abstract classes in the public api because it
makes it can make it a pain to write unit tests for the interface.
Because java 7 doesn't support default or static methods in the
interface, I had to add the ugly FileCaches class, but c'est la vie.