Catch InterruptedException in TimedPromise
I noticed that on Linux, the tests sometimes printed the stack trace of
the InterruptedException that can be thrown here. Note that I needed to
catch this here because scala doesn't have checked exceptions. If this
were implemented in java, it wouldn't have compiled because ultimately
it was part of a Runnable which does not have a throws in the signature
of run.