Refactor NioPathWatcher
The structure of NioPathWatcher and NioPathWatcherImpl was flawed.
Because of the way that I was using inheritance to pass the handleEvent
and handleOverflow methods, I had to perform some strange contortions on
the scala.js side. This is simplified by adding a new class
NioPathWatcherService that does not descend from NioPathWatcher and
having the NioPathWatcher make one as needed. It can then pass in the
handleEvent and handleOverflow methods as a Consumers to the
NioPathWatcherService, simplifying the code.

I also removed the options from the NioPathWatcher. There was no option
that the user would actually be able to control with the public api. The
one option that existed controlled whether the PathWatcher should
control its own DirectoryRegistry. It was straightforward to just pass
in a null DirectoryRegistry in the case where the PathWatcher is
managing its own registry*.

* straightforward except that it did lead to surprise
NullPointerExceptions on scala.js due to the coupling that had been
present previously.