Fix unregister bug
The NioFileCacheTest would periodically fail and I noticed that, indeed,
sometimes events were detected after the path was unmonitored. It turns
out that because I omitted a full directory cleanup in unregister, it
was possible to leave behind watch keys that were no longer needed (and
could trigger spurious events).