Refactor Entries
I didn't like the fact that I increased the object size by including the
IOException in the Entry so I refactored things so that Entry is now an
interface with two concrete implementations. By reworking it, I also
found areas in which it was possible for an unhandled IOException to
occur while creating an entry. To work around this, I replaced those
calls with calls to Entries.get which automatically handles the
IOExceptions.

This change ended up cascading because I ran into some issues with
Either that required some refactoring. Java and scala handle type
variance differently and in order to make things work on both platforms,
I had to move the left and right methods out of the Either class and
added leftProjection and rightProjection static methods. This is because
java handles co and contravariance of method parameters but not class
type parameters. This also required re-writing Either.scala to make the
Either class covariant in both type parameters. I also factored out the
common interface TypedPath from QuickFile and Entry.