Refactor Entries
I didn't like the fact that I increased the object size by including the
IOException in the Entry so I refactored things so that Entry is now an
interface with two concrete implementations. By reworking it, I also
found areas in which it was possible for an unhandled IOException to
occur while creating an entry. To work around this, I replaced those
calls with calls to Entries.get which automatically handles the
IOExceptions.