Add repository interface
The motivation for this change is that I want to add an implementation
of FileCache that does not actually cache the directory entries. My
suspicion is that there are still some residual bugs in managing the
directory cache that may cause the FileCache to get irrevocably out of
sync with the file system. I haven't seen it often, but it does seem to
occasionally happen. Since this would be very annoying to sbt users (the
current target audience of the library), I wanted to make it
configurable whether or not the FileCache actually caches the values.
There is also a significant memory overhead to caching the that works
out to roughtly 500 bytes per cached path. For sufficiently full
directories, this may cause issues. This made it clear that my
abstraction was not quite right since FileCache may not actually cache
anything. Moreover, there are shared interfaces between PathWatcher,
FileCache and Directory that I want to formalize.

The repository interface is very similar to QuickLister (which I may end
up renaming as well) and just provides a method for listing the contents
of a Path. I also add the DataRepository interface which has a list
method that returns Directory.Entry instances rather than path
instances. Finally, I add the DirectoryRepository interface which adds a
list method without a prefix.

Note that the Repository interface does not specify whether or not the
values are cached. It just provides an interface for fetching the files
(and possibly their entry value).