Refactor the project
This change should not add any features, as far as I am aware, but it
does end up touching the majority of source files in the project because
I renamed/refactored so many classes.

In the course of preparing a 1.4.0 release, I was examining the javadocs
and noticed that there were a number of overlapping concepts that did
not share a common interface. For example, QuickLister had a similar api
to Directory, but they did not share an interface. This led to more
public classes in the api than I preferred. To improve things, I
converted a number of classes to interface + implementation class and
re-worked things so that there were a few basic interfaces that build on
each other to create the full on interface. In general, I tried to stick
to the pattern of creating an interface and creating package private
implementations that can be instantiated using static methods on a `s`
class, e.g. `FileTreeViews` provides instances of `FileTreeView`.

An advantage of this approach is that it makes it very easy for a user
of the library to extend or modify the default implementations using
delegation. When using abstract classes, private state may leak/be
duplicated or it may not be practical to instantiate an instance.
Interfaces are trivial to implement using an IDE or even the error
messages printed by the scala compiler (which provides stubs). It also
gives me a lot of freedom moving forward to tweak the implementation
without breaking binary compatibility. At this point, I'm reasonably
happy with the interface and am optimistic that it won't be necessary to
make breaking changes in the future.

I also removed most of the redundant methods that added syntactic sugar
via defaults. It is easy for a user to add that syntactic sugar on their
own and I minimize the binary compatibility footprint by not providing
these apis out of the box.

There were numerous bugfixes that are incorporated in this change. For
the most part, however, I moved code around without changing much of the
actual method implementations.

As part of these changes I tried to be a bit more consistent with the
naming. The main interfaces are FileTreeView, FileTreeDataView,
Observable, ObservableCache, PathWatcher and FileTreeRepository. The
last of which incorporates all of the preceding interfaces.