Make it possible to remove shutdown hooks
I noticed when running the AllTests task that the heap size would grow
without bound. This is because the shutdown hooks to remove the
temporary files were holding references to the file names of the files
that were generally safely deleted by the test prior to shutdown. To fix
this memory leak, I make it possible to remove a shutdown hook. I also
removed some dead code.