Improve native library garbage collection
My previous approach to cleaning up the temporary shared libraries that
are extracted by the NativeLoader can leak files if the jvm is abruptly
shutdown. To fix this, on non-windows platforms, when we extract the
shared library to a temp file, we also create a corresponding file with
`.pid` appended to the file name. On startup, the NativeLoader will look
for the `.pid` files and check if there is a currently running process
with that pid. If not, it will delete the shared library file and
the .pid file. The shutdown hook will also attempt to delete both files,
so in the common case of clean jvm shutdown, nothing should be left
behind. On windows, the problem is complicated by the fact that the
shared library cannot be deleted while it's in use and there is no way
to unload the shared library without resorting to using reflection to
access private jvm apis, which now triggers a warning on newer jvms.
Instead, on startup, the NativeLoader will just try and delete the
shared libraries. If another process is still using it, the os will not
allow the delete, which is fine. We'll just try again next time.