Add SymlinkFollowingPathWatcher
The previous implementation of PathWatcher did not follow symlinks. To
attain parity of the monitoring api with the FileTreeRepositoryImpl, I
add the SymlinkFollowingPathWatcher class, which adds the ability to
monitor the targets of symlinks that exist in a registered path, or one
of its children. I also add the ability to toggle whether or not the
FileTreeRepository implementation follows symlinks or not.

The motivation for this change is that there are cases in which a user
may not want the overhead of a full on FileTreeRepository (which must
cache a lot of data if it is monitoring paths that include many files),
but may still want the same semantics. I don't provide a
FileTreeRepository that does not cache the paths, but it should be very
straightforward to implement the FileTreeRepository interface with the
SymlinkFollowingPathWatcher and a simple FileTreeView.