Fix file cache symlink bug
The FileCacheSymlinkTest occasionally (but consistently) failed,
especially when run as part of AllTests. It turns out that the reason
was that sometimes an apple file system event fired for the symlink (for
reasons that I do not really understand given that in that version of
the test, the call to register was made after the link was created).
When this event was handled by the FileCacheDirectoryTree, it passed in
the event's TypedPath to the CachedDirectoryImpl update method. This was
a problem because the link ended up being treated as a directory when,
if follow links is false, it should be treated like a file. The fix was
just to add exactly that logic to handleEvent.

I also added a test that a symlink that is added is treated like a file,
which should trigger a similar condition to that described in the
preceding paragraph.