Add swoval namespace
I broke binary compatibility of the swoval_apple_file_system.* source
files. This can cause segfaults if an old version of the library is
loaded by one classloader and a new version by the other. The segfaults
are avoided by moving all of the code into a namespace. I also could
have renamed some of the offending structures, but this was easier.
Plus, I still have room to add com to the namespace if I makde future
changes.