Support sbt 1.2.x
Unfortunately, binary compatibility of sbt.internal is not guaranteed
and a parameter was added to an internal api that I was using. To
resolve this, I must use reflection to get a handle to the Load.reapply
method and fill in a default logger if the newer version is found.