Don't exclude symbolic links when followLinks is true
I discovered an issue on linux where when registering a directory that
was a symbolic link, no watch events were detected. This was because the
filter for the CachedDirectoryImpl explicitly excluded symbolic links.
This was problematic because the logic for adding symbolic links only
applied to symbolic links that were found within a registered directory,
not the registered directory itself. With this change, I can likely
change the logic for addSymlink to only apply to files, but I am not
sure how necessary that is at the moment.