Make minor improvement to SymlinkFollowingPathWatcher
I realized that it was a mistake to call PathWatchers.get in the
constructor to SymlinkFollowingPathWatcher because it could lead to
stack overflow if followLinks was set to true. I also realized that when
a directory is detected by the path watcher, it's necessary to query the
directory for any symlinks that may be contained within it. The file
cache already did this, but not the standalone symlink following path
watcher.