Adjust overlapping directory semantics
I discovered some edge cases that could occur if a base directory was
registered non-recursively and then one of its immediate children, say
foo,  is separately registered. When this would happen, the base
registration would effectively be extended to cover the child, up to the
child's max depth. This isn't correct, though, because that means that
the directory would also include the other children, up to the max depth
of foo. To fix this, I create the directories with a filter that allows
a directory to include only those files accepted by the filter.
Conveniently, I already had such a filter, the DirectoryRegistry.

I also had to clean up the list of directories in the
FileCacheDirectoryTree whenever there is an update, because it's
possible that two independently registered directories could have a new
overlap and we only want a single entry for the least specific path when
that occurs.

I also removed the package private cachedUpdatable static method from
FileTreeDataViews to keep that class smaller.