Rework Entry interface
I realized that it was somewhat of a pain to have Entry extend TypedPath
because equality would not work the way that I wanted it to. I had
worked around this in the past, but I think it's worth fixing here. It
does make the code marginally more verbose, but I think that it's more
clear what's going on. More importantly, the equality semantics are
easier to reason about.