Export js api
This commit adds a number of wrapper classes that are used to export a
subset of the file-tree-view public api to native javascript. Because
js apps will not have access to java.nio or the swoval classes that
aren't extracted, all of the apis are expressed in terms of native js
features, e.g. javascript functions rather than java functional
interfaces or javascript strings rather than java.nio.file.Path.