Improve PathWatcherTest
I noticed that sometimes the thunk in Defer would throw an uncaught
exception because it was modifying a file that had been deleted. This
implied that an event was being detected before the thunk had run, which
was not expected. To fix this, I reworked a few tests so that they only
fired an event when the file reached an expected state. I also added an
exception in Defer.apply so that if an exception is thrown, we can
figure out from which test it originated.