Manage file handles correctly
I discovered that sometimes on windows, if there were multiple threads
performing io on the same directory, that the process could hang. This
would specifically happen when trying to list a path, while
simultaneously creating it. The big problem was that in a number of
instances, the windows file handle returned by opendDir was not closed
which would block the create from happening. To fix this, I ensure that
we close the handle whenever there is an error and before we throw an
exception.

I also minimize the amount of time that we hold the handle by resolving
unknown file types only after reading all of the file names from the
directory. Reading the file attributes can be slow, so this can
potentially significantly lower the amount of time we hold the handle
for the directory.