Fix non-recursive directory bug
I introduced a bug in 464e33961afe73aa8611c605a2300c042f56bad9 where any
directory that was added with a non negative one valued maxDepth
parameter would be created with a maxDepth of Integer.MAX_VALUE. It
would not actually include the recursive files because of the filter
that applies the DirectoryRegister.maxDepth parameter so we ended up in
an inconsistent state where the CachedDirectory purported to have depth
Integer.MAX_VALUE, but, in practice, had depth 0. When this directory
was the base directory of an sbt project, this caused problems because
we'd only see the files in the base directory and none of the children.

The new regression test fails if run without the FileCacheDirectoryTree
change.