Fix default file tree view semantics
This commit makes the default file tree view have the same semantics as
the default cached file tree view. Prior to this, there were two issues:

1) When the depth was specified as -1, it was effectively treated as 0.
   This made it impossible to get the TypedPath for a specific file
   through the file tree view.
2) Don't throw automatically re-throw exceptions that are encountered
   when filling the results. It may be expected that, say, an
   AccessDeniedException is thrown when recursively listing a directory.
   In most cases, we'd want to simply ignore the privileged file and
   continue filling the results. This is now configurable via a flag in
   FileTreeViews.getDefault. I did change the default to ignore
   exceptions. I figure it doesn't really matter because the only
   downstream consumer is sbt and no version has been released using
   this api.

Also, there was an issue where in the cached directory case, calling
`view.list(path, 0, AllPass)` would return the path itself it were
actually a file. This was not intentional, so I fixed that.