Fix off by one error with registration order
I found an obscure bug in sbt where when using a FileTreeRepository, it
was possible for events to be dropped. Suppose that you have base
directory foo and subdirectoyr foo/bar. Now suppose that we first
register foo/bar with depth 0 and then register foo with depth zero. The
cleanupDirectories callback would actually remove the registration for
foo/bar even though foo was not registered with depth 1, which would be
necessary for it to actually cover foo/bar if foo/bar had been
registered with depth zero.

Off by one errors are never fun, but at least I added a regression test