Add support for relative paths
It was reported by @scf37 that relative paths do not work for monitor
registration. In this commit, I change every implementation of
PathWatcher.register and PathWatcher.unregister to convert the input
path to an absolute path, if necessary. I also add a basic test that
relative work for both a PathWatcher and a FileTreeRepository.