Don't rescan directories by default
I noticed that there was a huge amount of io contention when using the
file cache with windows. Ultimately I tracked this down to windows
generating a good number of directory events which would cause a full
poll of said directory. This is generally unnecessary since we're
already monitoring the directory and incremental updates should work.