Don't monitor the root directory by default
It is generally a bad idea to monitor the root directory. Long term, I
should update the api to provide a set of valid root directories and
assert that the rawPath starts with one of them. For now, I'm just going
to avoid going  all the way up to the root..