Rescan on overflow
After 7c1c5db95f6d955881f8c482ea3fe4b60243b526, I noticed that the
NioFileCacheOverflowTest would sometimes fail on osx. It turns out that
the problem was that we were relying on the fact that update would
rescan the directory in the overflow scenario. It was almost trivial to
restore this behavior by changing FileCacheDirectoryTree.handleEvent to
take an Event rather than a typed Path.