Add logger to swoval
It is quite difficult to debug if the swoval libray is working without
any log messages. To remedy this, I add a new simple logging framework
that can be turned on by running the program with -Dswoval.debug=true.
Alternatively, the user can provide a full qualified class name for a
class that implements com.swoval.logging.Logger and has a no argument
constructor.