Rework tests
This commit adds a number of helper extension methods that can be used
to make the tests more convenient to write and robust to ephemeral io
exceptions (which can be a problem on windows). I added a `path`
extension method to FileTreeDataViews.Entry and PathWatchers.Event so
that we can avoid calling getTypedPath.getPath all over the place. I
also added a number of extendsion methods to Path that mimic the apis
available in Files, but automatically retry a few times if the test
fail. Finally, I ensured that all of the tests that wait for an event
actually ensure that the event is for the correct path before
proceeding. Previously there were cases where say a directory would
trigger an event before a file was created and it would cause a
transient test failure.