Update FileCacheOverflowTest
This test would sometimes fail when run as part of the AllTests suite. I
believe that it wasn't explicit enough about how it detected callback
events and decremented the various latches. Now everything is quite
explicit so that we can ensure that each file will only decrement the
latch once per event kind.

I also shuffle the order of events to make sure that we aren't relying on some
specific access pattern.