Improve AllTest
I've noticed that the test success rate is varied with the number of
threads that are running the tests. To try and test a wide variety of
threading conditions, I randomize the group size with each run. I bias
the group size towards 1 so that more threads are used more of the time,
but larger values approaching n (where n is the total number of tests to run)
may still be generated.