Refactor logging
It was very painful to track down bugs in tests because the logger was
specified globally. I rework all of the tests so that they each have
their own private logger that can be dumped if the test fails.

Right now the logging is global per test suite. Long term I'd like to
make it local to each spec, but that will require a bit of work.