Fix watch flakiness bug
There was a bug that would occasionally occur on appveyor where the
NioPathWatcherTest.registeredFiles.replace test would fail to detect a
file event. After much debugging, I finally realized that the problem
was that in the handleEvent call back of NioPathWatcher, we did not
correctly handle Delete events. Normally I stat the file to calculate
the event type. That was a mistake here because if the
NioPathWatcherService says a directory was deleted, then we must remove
that watch. The service doesn't manged the life cycle of its watched
directories so it is incumbent on the NioPathWatcher to correctly remove
them.

I also ended up refactoring this file quite a bit while trying to track
down the bug. Though they do make this change harder to parse as a diff,
I'm tired.